Updating separating utility functions into files
